+++ src/features/assistantList/components/NameTag.tsx
@@ -1,18 +1,35 @@
 import styled from "styled-components/native";
 
-export const NameTagBox = styled.View`
+type NameTagProps = {
+  name: string;
+  tag1: string;
+  tag2: string;
+};
+
+export default function NameTag({ name, tag1, tag2 }: NameTagProps) {
+  return (
+    <NameTagBox>
+      <NameText>{name}</NameText>
+      <TagText>
+        {tag1} {tag2}
+      </TagText>
+    </NameTagBox>
+  );
+}
+
+const NameTagBox = styled.View`
   justify-content: flex-start;
   align-items: flex-start;
   gap: ${({ theme }) => theme.gap.base};
 `;
 
-export const NameText = styled.Text`
+const NameText = styled.Text`
   font-family: ${({ theme }) => theme.fontFamily.nsExtraBold};
   font-size: ${({ theme }) => theme.fontSize.xl};
   color: ${({ theme }) => theme.colors.white};
 `;
 
-export const TagText = styled.Text`
+const TagText = styled.Text`
   font-family: ${({ theme }) => theme.fontFamily.nsBold};
   font-size: ${({ theme }) => theme.fontSize.xs};
   color: ${({ theme }) => theme.colors.white};
