+++ src/features/assistantList/components/ProfileImage.tsx
@@ -2,7 +2,19 @@ import styled from "styled-components/native";
 import CachedImage from "@/src/components/ui/CachedImage";
 import responsiveToPx from "@/src/utils/responsiveToPx";
 
-export const ImageBox = styled.View`
+type ProfileImageProps = {
+  url: string;
+};
+
+export default function ProfileImage({ url }: ProfileImageProps) {
+  return (
+    <ImageBox>
+      <Image src={url} />
+    </ImageBox>
+  );
+}
+
+const ImageBox = styled.View`
   width: ${responsiveToPx("300px")};
   height: ${responsiveToPx("300px")};
   background-color: ${({ theme }) => theme.colors.whiteBlue};
@@ -11,7 +23,7 @@ export const ImageBox = styled.View`
   align-items: center;
 `;
 
-export const Image = styled(CachedImage)`
+const Image = styled(CachedImage)`
   width: ${responsiveToPx("295px")};
   height: ${responsiveToPx("295px")};
   border-radius: 5px;
