+++ src/features/assistantList/containers/AssistantCardContainer.tsx
@@ -1,0 +1,10 @@
+import AssistantCard from "../components/AssistantCard";
+import type { TAssistantProfile } from "../types/assistantListTypes";
+
+type AssistantCardContainerProps = {
+  item: TAssistantProfile;
+};
+
+export default function AssistantCardContainer({ item }: AssistantCardContainerProps) {
+  return <AssistantCard item={item} onPressAiCard={() => {}} />;
+}
