+++ src/types/aiProfileTypes.ts
@@ -9,24 +9,6 @@ export type AiProfileMakeAnswers = {
   q6: string;
 };
 
-export type AiProfile = {
-  aiProfileId: number;
-  profileName: string;
-  imageUrl: string;
-  hashTag1: string;
-  hashTag2: string;
-  feature1: string;
-  feature2: string;
-  feature3: string;
-  createdAt: string;
-};
-
-export type AiProfileListWithGenerateAiTrigger = AiProfile | { isGenerateButton: boolean };
-
-export type AiProfileListResult = SuccessDTO & {
-  result: AiProfile[];
-};
-
 export type GetMakeAssistantQuestionResult = SuccessDTO & {
   result: AiProfileMakeAnswers & {
     createdAt: string;
